requirements.txt
openai-whisper
ffmpeg-python
yt-dlp
requests
Pillow
